+++ src/components/Popover/tests/index.js
@@ -30,6 +30,7 @@ describe('components/Popover', () => {
     );
     expect(component.children().attr('style')).toContain('display:block');
     expect(component.children().attr('style')).toContain('opacity:1');
+    expect(component.children().attr('style')).toContain('position:initial');
   });
 
   it('should have a properly set test section', () => {